Refactor data_vault and data_vault_multihead to share setting definitions
Review: @ejeffrey
